% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{My hello World function}
\usage{
hello(x)
}
\arguments{
\item{x}{the name of the person to say hi to}
}
\value{
the output from \code{\link{print}}
}
\description{
My hello World function
}
\examples{
hello("Delia")
\dontrun{
hello("Steve")
}

}
